rats-kafka-producer/pyproject.toml
[project]
name = "rats_kafka_producer"
version = "0.1.0"
description = "A Kafka producer for RATS job scraper."
requires-python = ">=3.12, <3.13"
readme = "README.md"
dependencies = [
    # Data Processing
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "loguru>=0.7.0", # Better logging
    "tenacity>=8.2.0", # Retry logic
    "pydantic>=2.5.0",
    "pydantic-settings>=2.12.0",
    "typer>=0.12.3",
    # JobSpy
    "python-jobspy>=1.1.82",
    # Kafka
    "confluent-kafka[avro,schemaregistry]>=2.13.0",
    "datacontract-cli[all,kafka,s3]>=0.10.16",
]

[project.scripts]
rats-kafka-producer = "rats_kafka_producer.cli:app"

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rats_kafka_producer"]

[tool.hatch.build.targets.sdist]
packages = ["src/rats_kafka_producer"]
